Fidelating
To quickly backup and restore data (especially large amounts of data), refer to the following recommendations:

Keep the exported data file as small as possible. It is recommended to use the -F option of Dumpling to set the file size. If you use MilevaDB Anyon to restore data, it is recommended that you set the value of -F to 256m.
If some of the exported tables have many rows, you can enable concurrency in the table by setting the -r option.